#913 move colored module from node.py to utils.py
