+++ src/psyclone/psyir/nodes/node.py
@@ -17,32 +17,10 @@ import graphviz
 
 from psyclone.core import VariablesAccessMap
 from psyclone.errors import GenerationError, InternalError
+from psyclone.utils import colored
 
 if TYPE_CHECKING:
     from psyclone.psyir.symbols import Symbol
-
-# We use the termcolor module (if available) to enable us to produce
-# coloured, textual representations of Invoke schedules. If it's not
-# available then we don't use colour.
-try:
-    # pylint disable=import-outside-toplevel
-    from termcolor import colored
-except ImportError:
-    # We don't have the termcolor package available (e.g. installing from)
-    # Spack) so provide alternative routine
-    def colored(text, _):
-        '''
-        Returns the supplied text argument unchanged. This is a swap-in
-        replacement for when termcolor.colored is not available.
-
-        :param str text: text to return.
-        :param _: fake argument, only required to match interface
-                  provided by termcolor.colored.
-
-        :returns: the supplied text, unchanged.
-        :rtype: str
-        '''
-        return text
 
 
 class ChildrenList(list):
